fix(ci): repair the generated-source template, 32-bit builds, and the ledger
CI has been red since the subpackage extraction. Four distinct causes.

The decoder cursor template was never updated when that commit exported
isDigit, pointerAt, and byteSourceOf. Its two generated files were hand-edited
instead, so the checked-in output was correct and its source was stale, and
regenerating replaced working code with references to identifiers that no
longer exist. The generated files now reproduce byte-for-byte.

Two constants overflowed a 32-bit int, both introduced during this work:
index_tree.go compared a length against int(^uint32(0)), and query's whole-number
narrowing compared a uint against an untyped 1<<62. Both are now widened to
uint64 before comparing. linux/386, linux/arm, and s390x all build clean.

staticcheck reported a dead store in the exact index. It is genuinely redundant
rather than a missing check: the packed iterator returns its zero value on a
lookup miss and reports exhaustion immediately, so the priming next() already
folds "absent from the base" into the flag that is kept. Discarded explicitly
with a note, so the next reader does not restore it as a missing guard.

The test-ownership ledger still listed thirty-nine test files at their
pre-subpackage paths and was missing five added since. Entries are relocated
against the recorded renames so each keeps its original contract, and the
checker is untouched.

Finally, the direct indexed count test asserted both correctness and zero
allocations under one name. The checkptr and race jobs skip tests whose name
matches Alloc, because pointer instrumentation allocates on its own account —
so that test failed under checkptr, and simply renaming it would have dropped
its correctness assertions from two jobs. The allocation assertion moves to its
own suitably named test and the correctness assertions stay put, sharing one
fixture.

+++ internal/storeio/index_tree.go
@@ -56,7 +56,9 @@ func (p *IndexTreeProbe) appendEntry(view IndexDirectoryView, entry IndexDirecto
 			}
 		}
 		if span.Length == 0 {
-			if len(p.Certificates) > int(^uint32(0))-len(certificate) {
+			// Compared in uint64: the bound is a uint32 maximum, which does
+			// not fit an int on a 32-bit platform and would not compile there.
+			if uint64(len(p.Certificates)) > uint64(^uint32(0))-uint64(len(certificate)) {
 				return ErrIndexProbeCertificates
 			}
 			span = CertSpan{Offset: uint32(len(p.Certificates)), Length: uint16(len(certificate))}
